adecuaciones

+++ DragonGolfBackEnd/Controllers/Apuestas/TeamNassau/CrearDetalleApuestaTeamController.cs
@@ -43,7 +43,7 @@ namespace DragonGolfBackEnd.Controllers
             public float BetD_Medal { get; set; }
             public int BetD_AutoPress { get; set; }
             public int BetD_ManuallyOverrideAdv { get; set; }
-            public int BetD_AdvStrokers { get; set; }
+            public double BetD_AdvStrokers { get; set; }
             public string TypeHandicap { get; set; }
         }
 
@@ -71,7 +71,7 @@ namespace DragonGolfBackEnd.Controllers
                 comando.Parameters.Add("@BetD_Medal", SqlDbType.Float);
                 comando.Parameters.Add("@BetD_AutoPress", SqlDbType.Int);
                 comando.Parameters.Add("@BetD_ManuallyOverrideAdv", SqlDbType.Int);
-                comando.Parameters.Add("@BetD_AdvStrokers", SqlDbType.Int);
+                comando.Parameters.Add("@BetD_AdvStrokers", SqlDbType.Float);
                 comando.Parameters.Add("@TypeHandicap", SqlDbType.VarChar);
